+++ src/Neo.SmartContract.Framework/Native/AndCondition.cs
@@ -10,9 +10,8 @@
 
 namespace Neo.SmartContract.Framework.Native
 {
-    public class WitnessRule
+    public class AndCondition : WitnessCondition
     {
-        public WitnessRuleAction Action;
-        public WitnessCondition Condition;
+        public WitnessCondition[] Expressions;
     }
 }
